requirements.txt
azure-core==1.29.7
azure-identity==1.15.0
azure-search-documents==11.4.0
semantic-kernel==0.4.5.dev0
openai==1.9.0
langchain==0.1.1
langchain-openai==0.0.3
tiktoken==0.5.2
python-dotenv==1.0.0
requests==2.31.0
unstructured==0.12.2
markdown==3.5.2
qdrant-client==1.7.1
chromadb==0.4.22
yfinance==0.2.36
langchain-community
pymongo
